refactor: split victory overlay, empire rail, and civilian draft parts (#3878)
Extract VictoryPanel decorations, empire rail button chrome, and civilian
draft projection helpers into library parts for Phase 3 flame modularization;
public APIs and behavior unchanged.
