+++ app/lib/features/game/flame/map_state/game_map_area_civilian_draft_projection.dart
@@ -4,6 +4,8 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart' as ct_models;
 import 'package:colonizethis_map/colonizethis_map.dart';
 
+part 'game_map_area_civilian_draft_projection_helpers.dart';
+
 /// Civilian-marker draft projection for the human player.
 ///
 /// Extracted from `GameMapAreaStateLogic` (#2575 work item 11) so the
@@ -109,148 +111,17 @@ class GameMapAreaCivilianDraftProjection {
           );
     }
     if (projectedByTile.isEmpty) {
-      return RegionMapViewData(
-        regionId: region.regionId,
-        width: region.width,
-        height: region.height,
-        cellSize: region.cellSize,
-        cells: region.cells,
-        capitalMarkers: region.capitalMarkers,
-        portMarkers: region.portMarkers,
-        factionColors: region.factionColors,
-        greatPowerFactionIds: region.greatPowerFactionIds,
-        terrainColors: region.terrainColors,
-        unitMarkers: region.unitMarkers,
-        civilianTileMarkers: const [],
-        fleetTileMarkers: region.fleetTileMarkers,
-        warpMarkers: region.warpMarkers,
-        townMarkers: region.townMarkers,
-        provinceUnitPresenceByProvinceId:
-            region.provinceUnitPresenceByProvinceId,
-        provincePoliticalOwnerByPrefixedProvinceId:
-            region.provincePoliticalOwnerByPrefixedProvinceId,
-        seaZoneDisplayNameByPrefixedId: region.seaZoneDisplayNameByPrefixedId,
-      );
+      return _regionWithEmptyCivilianMarkers(region);
     }
 
-    final projectedMarkers = <CivilianTileMarkerView>[];
-    for (final entry in projectedByTile.entries) {
-      final tileKey = entry.key;
-      final units = entry.value.toList()
-        ..sort((a, b) {
-          final p = _civilianIconPriorityForType(
-            a.unitType,
-          ).compareTo(_civilianIconPriorityForType(b.unitType));
-          if (p != 0) return p;
-          return a.unitId.compareTo(b.unitId);
-        });
-      final parsed = tryParseTileKey(tileKey);
-      if (parsed == null) continue;
-      final x = parsed.x;
-      final y = parsed.y;
-      final representative = units.first;
-      final representativeIsAssigned =
-          representative.pendingTargetTileKey == tileKey ||
-          (representative.assignedTileKey == tileKey &&
-              representative.status == ct_models.UnitStatus.working);
-      final applyCivilianRevealHalo = units.any((u) {
-        final isAssignedToTile =
-            u.pendingTargetTileKey == tileKey ||
-            (u.assignedTileKey == tileKey &&
-                u.status == ct_models.UnitStatus.working);
-        if (!isAssignedToTile) return false;
-        return visibilityByTile[tileKey] == VisibilityLevel.fogged.name;
-      });
-      projectedMarkers.add(
-        CivilianTileMarkerView(
-          tileKey: tileKey,
-          x: x,
-          y: y,
-          localProvinceId: parsed.provinceLocalId,
-          unitIds: units.map((u) => u.unitId).toList(),
-          unitTypes: {for (final u in units) u.unitId: u.unitType},
-          representativeUnitType: representative.unitType,
-          stackCount: units.length,
-          representativeIsAssigned: representativeIsAssigned,
-          applyCivilianRevealHalo: applyCivilianRevealHalo,
-        ),
-      );
-    }
+    final projectedMarkers = _buildProjectedCivilianMarkers(
+      projectedByTile: projectedByTile,
+      visibilityByTile: visibilityByTile,
+    );
 
-    projectedMarkers.sort((a, b) {
-      final yc = a.y.compareTo(b.y);
-      if (yc != 0) return yc;
-      final xc = a.x.compareTo(b.x);
-      if (xc != 0) return xc;
-      return a.tileKey.compareTo(b.tileKey);
-    });
-
-    return RegionMapViewData(
-      regionId: region.regionId,
-      width: region.width,
-      height: region.height,
-      cellSize: region.cellSize,
-      cells: region.cells,
-      capitalMarkers: region.capitalMarkers,
-      portMarkers: region.portMarkers,
-      factionColors: region.factionColors,
-      greatPowerFactionIds: region.greatPowerFactionIds,
-      terrainColors: region.terrainColors,
-      unitMarkers: region.unitMarkers,
-      civilianTileMarkers: projectedMarkers,
-      fleetTileMarkers: region.fleetTileMarkers,
-      warpMarkers: region.warpMarkers,
-      townMarkers: region.townMarkers,
-      provinceUnitPresenceByProvinceId: region.provinceUnitPresenceByProvinceId,
-      provincePoliticalOwnerByPrefixedProvinceId:
-          region.provincePoliticalOwnerByPrefixedProvinceId,
-      seaZoneDisplayNameByPrefixedId: region.seaZoneDisplayNameByPrefixedId,
+    return _regionWithProjectedCivilianMarkers(
+      region: region,
+      projectedMarkers: projectedMarkers,
     );
   }
-}
-
-bool _isCivilianUnitType(String unitType) {
-  final role = unitRoleForType(unitType);
-  if (role == null) return false;
-  return role != UnitRole.military && role != UnitRole.naval;
-}
-
-String _normalizeCivilianTypeForPriority(String type) {
-  return type.toLowerCase().replaceAll(RegExp(r'[\s_\-]'), '');
-}
-
-int _civilianIconPriorityForType(String type) {
-  final normalized = _normalizeCivilianTypeForPriority(type);
-  switch (normalized) {
-    case 'builder':
-      return 0;
-    case 'engineer':
-      return 1;
-    case 'railbuilder':
-      return 2;
-    case 'explorer':
-      return 3;
-    case 'merchant':
-      return 4;
-    case 'spy':
-      return 5;
-    default:
-      return 999;
-  }
-}
-
-class _ProjectedCivilianUnit {
-  const _ProjectedCivilianUnit({
-    required this.unitId,
-    required this.unitType,
-    required this.pendingTargetTileKey,
-    required this.assignedTileKey,
-    required this.status,
-  });
-
-  final String unitId;
-  final String unitType;
-  final String? pendingTargetTileKey;
-  final String? assignedTileKey;
-  final ct_models.UnitStatus status;
 }
